tools/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth
google-api-core<2.12
